[mv-button-reorg] Split tests
